requirements.txt
# pip freeze --isolated > requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
altgraph==0.17.2
annotated-types==0.7.0
anyio==4.8.0
async-timeout==4.0.3
attrs==24.3.0
certifi==2024.12.14
charset-normalizer==3.4.1
dataclasses-json==0.6.7
exceptiongroup==1.2.2
faiss-cpu
frozenlist==1.5.0
future==0.18.3
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.14
langchain-community==0.3.14
langchain-core==0.3.29
langchain-text-splitters==0.3.5
langchain_huggingface
langchain_openai
langchain_community
langsmith==0.2.10
macholib==1.15.2
marshmallow==3.25.1
multidict==6.1.0
mypy-extensions==1.0.0
nltk>=3.8.1
networkx>=2.8.0
numpy==1.26.4
openpyxl
orjson==3.10.14
pandas>=1.4.0
packaging==24.2
propcache==0.2.1
pydantic==2.10.5
pydantic-settings==2.7.1
pydantic_core==2.27.2
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
sentence-transformers
six>=1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.37
tenacity==9.0.0
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.2
unstructured>=0.6.7
xlrd>=2.0.1
yarl==1.18.3
